Move encoded_bytes_stream into tests infra
