+++ tests/infra/__init__.py
@@ -1,4 +1,8 @@
+import sys
 import warnings
+import contextlib
+
+from io import TextIOWrapper, BytesIO
 
 
 class SuppressWarningsMixin:
@@ -7,3 +11,11 @@ class SuppressWarningsMixin:
         warnings.filterwarnings("ignore", "Your application has authenticated using end user credentials")
         # Suppress unclosed socket warnings
         warnings.simplefilter("ignore", ResourceWarning)
+
+@contextlib.contextmanager
+def encoded_bytes_stream():
+    old_stdout = sys.stdout
+    sys.stdout = TextIOWrapper(BytesIO(), sys.stdout.encoding)
+    yield
+    sys.stdout.close()
+    sys.stdout = old_stdout
