+++ tests/test_terra_notebook_utils_cli.py
@@ -21,7 +21,7 @@ import google_crc32c
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from tests import config, encoded_bytes_stream
+from tests import config
 from tests.infra.testmode import testmode
 from terra_notebook_utils import gs, WORKSPACE_NAME, WORKSPACE_GOOGLE_PROJECT, WORKSPACE_BUCKET
 from terra_notebook_utils.cli import Config
@@ -32,7 +32,7 @@ import terra_notebook_utils.cli.workspace
 import terra_notebook_utils.cli.profile
 import terra_notebook_utils.cli.drs
 import terra_notebook_utils.cli.table
-from tests.infra import SuppressWarningsMixin
+from tests.infra import SuppressWarningsMixin, encoded_bytes_stream
 
 
 @testmode("workspace_access")
